requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
pydantic-settings==2.6.0
openai==1.54.0
anthropic==0.40.0
redis==5.2.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
pyyaml==6.0.2
structlog==24.4.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.17
prometheus-client==0.21.0
mcp==1.6.0
pytest==8.3.0
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.0